generator/requirements.txt
google-api-python-client==1.7.11
pybars3==0.9.6
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
coverage==4.5.4